+++ decidim-core/config/routes.rb
@@ -47,6 +47,12 @@ Decidim::Core::Engine.routes.draw do
     end
     resource :notifications_settings, only: [:show, :update], controller: "notifications_settings"
     resources :own_user_groups, only: [:index]
+    resource :data_portability, only: [:show], controller: "data_portability" do
+      member do
+        post :export
+        get :download_file
+      end
+    end
   end
 
   resources :profiles, only: [:show], param: :nickname